Move repo logic to a data source
